Support replies and reactions
All text message events now include two extra fields - `reply_id` and
`emoji`.

These correspond to the same fields in the `Data` protobuf and can be
used to determine whether the message is a reply or a reaction.

The three service actions that send messages (Send Text, Broadcast
Channel Message and Send Direct Message) now accept an `emoji`
parameter, which will be reflected in the transmitted packet. This can
be used together with the existing `reply_id` parameter to send
reactions.

I've decided to also simplify some of the existing `reply_id` code. We
don't need an explicit `None` case, since zero already has exactly that
meaning. This makes some of the code simpler and some of the events more
predictable - `reply_id` now has exactly one value indicating that a
message is not a reply (i.e. zero) and not two (i.e. both zero and
`None`).

I've also altered the descriptions for the `reply_id` field. They're now
more in-line with the Meshtastic protobufs.

Finally, I've changed the `reply_id` selectors to be templates, since
you'll definitely be putting a template inside that field.